Improve bibliography modal UI
