+++ src/scss/components/_citation-options.scss
@@ -1,0 +1,22 @@
+.citation-options {
+	@include touch-or-bp-down(sm) {
+		margin-bottom: $form-group-margin-touch;
+	}
+
+	@include mouse-and-bp-up(md) {
+		position: sticky;
+		top: 0;
+		width: 100%;
+		background-color: $modal-content-bg;
+		padding-top: $form-group-margin;
+		padding-bottom: $form-group-margin;
+		border-bottom: $border-width solid $modal-border-color;
+		margin-bottom: $modal-inner-padding;
+	}
+
+	@include state(".modal-body:focus .citation-options") {
+		@include mouse-keyboard-and-bp-up(md) {
+			padding-top: $form-group-margin - $outline-width;
+		}
+	}
+}
